Added PHP 8.2 tests and PHP 8.5 compatibility

+++ src/Export/StreamingExcelExport.php
@@ -371,7 +371,7 @@ class StreamingExcelExport extends ExportAbstract
             if (is_int($result)) {
                 $result = (int) $result;
             } else {
-                $result = (double) $result;
+                $result = (float) $result;
             }
         } else {
             $result = $this->filterCsvInjection($result);
